New concepts maps for scenario 1 +2
disclaimer: just concept maps , very rough but enough for starting developing
map1 concept -  missing shops for merchants
map1  concept river - is a concept for how we can implement a cool water river. We'd need transitions for the wadi sands to water.  Creating a bigger scaled bridge should not be a problem. If you like how a the river could look like once the transitions are there, i would even have it in the middle of the map so the player has to travers it a bunch of times.
map 2 concept - is a draft for the scenario encounter with the desert guide.  enough for start developing that scenario as well